requirements.txt
chromadb>=1.4.1
datasets>=4.5.0
httpx>=0.28.1
langchain>=1.2.8
langchain-community>=0.4.1
pandas>=2.3.3
pypdf2>=3.0.1
python-docx>=1.2.0
sentence-transformers>=5.2.2
tiktoken>=0.12.0
torch>=2.10.0
pytest>=9.0.2
torchinfo>=1.8.0
transformers>=5.0.0
